updated controller and employee dao

+++ src/main/java/pl/piotrchlopicki/godelTask/utils/EmployeeRowMapper.java
@@ -10,8 +10,8 @@ public class EmployeeRowMapper implements RowMapper<Employee> {
     @Override
     public Employee mapRow(ResultSet rs, int arg1) throws SQLException {
         Employee employee = new Employee();
-        employee.setEmployeeId(rs.getInt("employee_id"));
-        employee.setDepartmentId(rs.getInt("department_id"));
+        employee.setEmployeeId(rs.getLong("employee_id"));
+        employee.setDepartmentId(rs.getLong("department_id"));
         employee.setFirstName(rs.getString("first_name"));
         employee.setLastName(rs.getString("last_name"));
         employee.setJobTitle(rs.getString("job_title"));
